requirements.txt
streamlit
tensorflow
Pillow
numpy
matplotlib
requests
gdown
